requirements_txt.txt
pyinstaller>=5.0.0
pillow>=9.0.0
tkinter